+++ jupyter/kernel/module.nix
@@ -2,7 +2,11 @@
 #
 # SPDX-License-Identifier: MPL-2.0 OR MIT
 
-{ name, config, lib, pkgs, ... }:
+# This module defines the “interface” of the kernel module, i.e. the options
+# that a kernel type’s author is expected to set to “register” their kernel
+# with Jupyter.
+
+{ lib, ... }:
 
 {
   options = {
@@ -16,6 +20,7 @@
     jupyterEnvPackages = lib.mkOption {
       type = lib.types.functionTo (lib.types.listOf lib.types.package);
       description = "Selector for extra Python packages to install into the Jupyter python environment";
+      internal = true;
       default = _: [];
       defaultText = lib.literalExpression ''_: []'';
     };
@@ -23,6 +28,7 @@
     jupyterExtensions = lib.mkOption {
       type = lib.types.listOf lib.types.package;
       description = "Packages containing Jupyter extensions to install. Be careful with package sets to ensure compatibility";
+      internal = true;
       default = [ ];
     };
   };